[Update]: Remove code duplication and redundancy after registering a global error handler

+++ backend/src/controller/userCtrl.mjs
@@ -8,28 +8,35 @@ import validateEmail from '../utils/validateEmail.mjs';
 // Define a getUserDetails function that returns a JSON
 // response with the specified user details
 
-const getUserDetails = async (req, res) => {
+const getUserDetails = async (req, res, next) => {
   const { _id } = req.user;
 
   try {
     const user = await User.findById(_id);
 
+    if (!user) {
+      return res.status(404).json({ success: false, message: 'User not found.' });
+    }
+
     res.status(200).json({ success: true, message: 'User details successfully retrieved.', details: user });
   } catch (error) {
-    console.error('Error retrieving user details:', error);
-    res.status(500).json({ success: false, message: 'Internal server error. Please try again later.' });
+    next(error);
   }
 };
 
 // Define updateUserDetails function that allows registered and authenticated
 // Users to update their details such as firstname, lastname, phone and/or password
 
-const updateUserDetails = async (req, res) => {
+const updateUserDetails = async (req, res, next) => {
   const { _id } = req.user;
   const { firstname, lastname, phone, previousPassword, password } = req.body;
 
   try {
     const user = await User.findById(_id);
+
+    if (!user) {
+      return res.status(404).json({ success: false, message: 'User not found.' });
+    }
 
     if (previousPassword) {
       const passwordMatch = await verifyData(user.password, previousPassword);
@@ -53,19 +60,22 @@ const updateUserDetails = async (req, res) => {
     await user.save();
     res.status(200).json({ success: true, message: 'User details successfully updated.', details: user });
   } catch (error) {
-    console.error('Error updating user details:', error);
-    res.status(500).json({ success: false, message: 'Internal server error. Please try again later.' });
+    next(error);
   }
 };
 
 // Define changeEmail function that allows registered and authenticated
 // Users to change their emails if need be
-const changeEmail = async (req, res) => {
+const changeEmail = async (req, res, next) => {
   const { _id } = req.user;
   const { email } = req.body;
 
   try {
     const user = await User.findById(_id);
+
+    if (!user) {
+      return res.status(404).json({ success: false, message: 'User not found.' });
+    }
 
     // Validate email
     await validateEmail(email);
@@ -84,8 +94,7 @@ const changeEmail = async (req, res) => {
 
     res.status(200).json({ success: true, message: 'Email successfully updated. Verify OTP.', details: user });
   } catch (error) {
-    console.error('Error changing email:', error);
-    res.status(500).json({ success: false, message: 'Internal server error. Please try again later.' });
+    next(error);
   }
 };
 
@@ -93,11 +102,15 @@ const changeEmail = async (req, res) => {
 // Users to suspend/deactivate their accounts
 // Can also be used by Admin
 
-const suspendAccount = async (req, res) => {
+const suspendAccount = async (req, res, next) => {
   const { _id } = req.user;
 
   try {
     const user = await User.findById(_id);
+
+    if (!user) {
+      return res.status(404).json({ success: false, message: 'User not found.' });
+    }
 
     user.active = false;
     await user.save();
@@ -106,8 +119,7 @@ const suspendAccount = async (req, res) => {
 
     res.status(200).json({ success: true, message: 'Account successfully deactivated.' });
   } catch (error) {
-    console.error('Error suspending account:', error);
-    res.status(500).json({ success: false, message: 'Internal server error. Please try again later.' });
+    next(error);
   }
 };
 
@@ -115,17 +127,22 @@ const suspendAccount = async (req, res) => {
 // Users to delete their accounts
 // Can also be used by Admin
 
-const deleteAccount = async (req, res) => {
+const deleteAccount = async (req, res, next) => {
   const { _id } = req.user;
 
   try {
+    const user = await User.findById(_id);
+
+    if (!user) {
+      return res.status(404).json({ success: false, message: 'User not found.' });
+    }
+
     await User.findByIdAndDelete(_id);
     req.session.destroy(); // Destroy session after deleting account
 
     res.status(200).json({ success: true, message: 'Account successfully deleted.' });
   } catch (error) {
-    console.error('Error deleting account:', error);
-    res.status(500).json({ success: false, message: 'Internal server error. Please try again later.' });
+    next(error);
   }
 };
